LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpSender-transport.https-expected.txt


PASS RTCRtpSender.transport is null when unconnected
PASS RTCRtpSender/receiver.transport has a value when connected
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy balanced
FAIL RTCRtpSender/receiver/SCTP transport at the right time, with bundle policy balanced assert_equals: Callee SCTP transport does not match: expected object "[object RTCDtlsTransport]" but got object "[object RTCDtlsTransport]"
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy max-bundle
FAIL RTCRtpSender/receiver/SCTP transport at the right time, with bundle policy max-bundle assert_equals: Callee SCTP transport does not match: expected object "[object RTCDtlsTransport]" but got object "[object RTCDtlsTransport]"
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy max-compat
FAIL RTCRtpSender/receiver/SCTP transport at the right time, with bundle policy max-compat assert_equals: Callee SCTP transport does not match: expected object "[object RTCDtlsTransport]" but got object "[object RTCDtlsTransport]"
